requirements.txt
aiolimiter==1.2.1
anyio==4.8.0
APScheduler==3.11.0
cachetools==5.5.1
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==22.3
sniffio==1.3.1
tornado==6.4.2
typing_extensions==4.12.2
tzlocal==5.3
